requirements.txt
fastapi
uvicorn
requests
openai
python-dotenv
langchain
openai
faiss-cpu
pypdf
sqlite-utils
